made changes to wishlists

+++ app/Http/Controllers/WishListController.php
@@ -7,6 +7,12 @@ use Illuminate\Http\Request;
 
 class WishListController extends Controller
 {
+    public function index()
+    {
+        $wishlists = auth()->user()->wishlists->load('product');
+        return view('wishlist', compact('wishlists'));
+    }
+
     public function store(Request $request)
     {
         $user = auth()->user();
@@ -27,15 +33,24 @@ class WishListController extends Controller
         return response(['available' => (int)$user->wishlists->contains('product_id', $product->id)], 200);
     }
 
-
+    
     public function destroy(Request $request)
     {
-        $data = $request->validate([
-            'product_id' => 'required|exists:products,id',
-        ]);
-
         $user = auth()->user();
-        $user->wishlists()->where('product_id', $data['product_id'])->delete();
-        return $user->wishlists->count();
+
+        if($request->ajax()){
+            $data = $request->validate([
+                'product_id' => 'required|exists:products,id',
+            ]);
+
+            $user->wishlists()->where('product_id', $data['product_id'])->delete();
+            return $user->wishlists->count();
+        }
+        $data = $request->validate([
+            'id' => 'required|exists:wish_lists,id',
+        ]);
+        $user->wishlists()->where('id', $data['id'])->delete();
+
+        return back();        
     }
 }
